Add symbolic links for authoritative name servers.
When Evolution fails to find a direct match for an email address
domain, it queries the domain's authoritative name server, and tries
to match some suffix of the authoritative name server's domain.

To make this work, add symbolic links for all relevant name server
domains back to the original Mozilla configuration file.